Ch. 17 §02: correct numbering and content for Listings 17-15 and 17-16
